+++ Firmament/src/utopia/firmament/model/stack/StackInsets.scala
@@ -1,10 +1,7 @@
 package utopia.firmament.model.stack
 
-import utopia.flow.operator.sign.Sign
-import utopia.flow.operator.sign.SignOrZero.Neutral
-import utopia.paradigm.enumeration.Axis.{X, Y}
-import utopia.paradigm.enumeration.{Alignment, Axis2D, Direction2D}
-import utopia.paradigm.shape.shape2d.insets.{Insets, ScalableSidesLike, SidedBuilder, SidesFactory}
+import utopia.paradigm.enumeration.Direction2D
+import utopia.paradigm.shape.shape2d.insets.{Insets, SidedBuilder, SidesFactory}
 
 object StackInsets extends SidesFactory[StackLength, StackInsets]
 {
@@ -49,7 +46,7 @@ object StackInsets extends SidesFactory[StackLength, StackInsets]
   * @since 2.2.2020, Reflection v1
   */
 case class StackInsets(sides: Map[Direction2D, StackLength])
-	extends ScalableSidesLike[StackLength, StackSize, StackInsets] with StackInsetsConvertible
+	extends StackInsetsLike[StackInsets] with StackInsetsConvertible
 {
 	// ATTRIBUTES	-----------------------
 	
@@ -64,57 +61,10 @@ case class StackInsets(sides: Map[Direction2D, StackLength])
 	  */
 	lazy val min = mapToInsets { _.min }
 	
-	override lazy val total: StackSize = StackSize(totalAlong(X), totalAlong(Y))
+	override lazy val total: StackSize = super.total
 	
 	
 	// COMPUTED	---------------------------
-	
-	/**
-	  * @return A copy of these insets where each side is marked with low priority
-	  */
-	def withLowPriority = StackInsets(Direction2D.values.map { d => d -> apply(d).lowPriority }.toMap)
-	
-	/**
-	  * @return A copy of these insets where no side has a maximum limit and all sides expand easily
-	  */
-	def expanding = map { _.expanding.noMax }
-	/**
-	  * @return A copy of these insets where horizontal sides have no maximum limit and expand easily
-	  */
-	def expandingHorizontally = expandingAlong(X)
-	/**
-	  * @return A copy of these insets where vertical sides have no maximum limit and expand easily
-	  */
-	def expandingVertically = expandingAlong(Y)
-	/**
-	  * @return A copy of these insets where the right side has no maximum limit and expands easily
-	  */
-	def expandingToRight = expandingTowards(Direction2D.Right)
-	/**
-	  * @return A copy of these insets where the left side has no maximum limit and expands easily
-	  */
-	def expandingToLeft = expandingTowards(Direction2D.Left)
-	/**
-	  * @return A copy of these insets where the top has no maximum limit and expands easily
-	  */
-	def expandingToTop = expandingTowards(Direction2D.Up)
-	/**
-	  * @return A copy of these insets where the bottom has no maximum limit and expands easily
-	  */
-	def expandingToBottom = expandingTowards(Direction2D.Down)
-	
-	/**
-	  * @return A copy of these insets with no minimum value (all minimum values at 0)
-	  */
-	def noMin = map { _.noMin }
-	/**
-	  * @return A copy of these insets with no maximum value
-	  */
-	def noMax = map { _.noMax }
-	/**
-	  * @return A copy of these insets with no minimum or maximum value
-	  */
-	def noLimits = map { _.noLimits }
 	
 	@deprecated("Please use .sides instead", "v1.5")
 	def amounts = sides
@@ -124,67 +74,13 @@ case class StackInsets(sides: Map[Direction2D, StackLength])
 	
 	override def self = this
 	
-	override protected def zeroLength = StackLength.fixedZero
-	
-	override protected def join(a: StackLength, b: StackLength): StackLength = a + b
-	override protected def subtract(from: StackLength, amount: StackLength): StackLength = from - amount
-	override protected def multiply(a: StackLength, multiplier: Double)  = a * multiplier
-	
-	override protected def withSides(sides: Map[Direction2D, StackLength]): StackInsets = StackInsets(sides)
-	
 	@deprecated("There's no need to call this method since 'this' already does this", "v2")
 	override def toInsets = this
 	
+	override protected def withSides(sides: Map[Direction2D, StackLength]): StackInsets = StackInsets(sides)
+	
 	
 	// OTHER	---------------------------
-	
-	/**
-	  * @param amount Length increase affecting each side of these insets
-	  * @return A copy of these insets with each side increased
-	  */
-	def +(amount: Double) = if (amount == 0) this else map { _ + amount }
-	/**
-	  * @param other A set of insets
-	  * @return A copy of these insets which have been increased by the other set of insets
-	  */
-	def +(other: Insets) =
-		withSides((sides.keySet ++ other.sides.keySet).map { dir => dir -> (apply(dir) + other(dir)) }.toMap)
-	/**
-	  * @param amount Length decrease affecting each side of these insets
-	  * @return A copy of these insets with each side decreased
-	  */
-	def -(amount: Double) = if (amount == 0) this else map { _ - amount }
-	/**
-	  * @param other A set of insets
-	  * @return A copy of these insets with the specified insets subtracted on all sides
-	  */
-	def -(other: Insets) = this + (-other)
-	
-	/**
-	  * @param other Another set of insets
-	  * @return Combination between these insets, which attempts to fulfill the conditions in both
-	  */
-	def &&(other: StackInsets) = mergeWith(other) { (a, b) =>
-		/*
-		val defaultMerge = a && b
-		// Preserves expanding insets, even when expansion is not listed in both
-		if (a.priority.expandsFirst || b.priority.expandsFirst)
-			defaultMerge.expanding
-		else
-			defaultMerge
-		 */
-		a && b
-	}
-	/**
-	  * @param other Another set of insets
-	  * @return A combination of these insets that selects the smaller applicable value
-	  */
-	def min(other: StackInsets) = mergeWith(other) { _ min _ }
-	/**
-	  * @param other Another set of insets
-	  * @return A combination of these insets that selects the larger applicable value
-	  */
-	def max(other: StackInsets) = mergeWith(other) { _ max _ }
 	
 	/**
 	  * Converts these stack insets to normal insets
@@ -192,38 +88,4 @@ case class StackInsets(sides: Map[Direction2D, StackLength])
 	  * @return A set of insets with mapped side values
 	  */
 	def mapToInsets(f: StackLength => Double) = Insets(sides.map { case (d, l) => d -> f(l) })
-	
-	/**
-	  * @param direction Target direction
-	  * @return A copy of these insets that expand to the specified direction, with no maximum value defined either.
-	  */
-	def expandingTowards(direction: Direction2D) = mapSide(direction) { _.expanding.noMax }
-	/**
-	  * @param axis Target axis
-	  * @return A copy of these insets that expand along the specified axis, with no maximum value defined either.
-	  */
-	def expandingAlong(axis: Axis2D) = mapAxis(axis) { _.expanding.noMax }
-	
-	/**
-	  * @param axis Target axis
-	  * @param alignment Contextual alignment
-	  * @return A copy of these insets which expand to the direction opposite to the aligned side. For example, if
-	  *         axis = X and alignment is Right, expands to Left. If alignment is center, expands to both directions.
-	  */
-	def expandingAccordingTo(axis: Axis2D, alignment: Alignment) = alignment(axis).direction match {
-		case preservedDirection: Sign => expandingTowards(axis.toDirection(preservedDirection.opposite))
-		case Neutral => expandingAlong(axis)
-	}
-	/**
-	  * @param alignment Contextual alignment
-	  * @return A copy of these insets that expand to the opposite direction of the aligned side. For example, if
-	  *         alignment is Right, expands to Left and if alignment is Center, expands to Left and Right.
-	  */
-	def expandingHorizontallyAccordingTo(alignment: Alignment) = expandingAccordingTo(X, alignment)
-	/**
-	  * @param alignment Contextual alignment
-	  * @return A copy of these insets that expand to the direction opposite to the aligned side. For example,
-	  *         if alignment is Top, expands to Bottom and if alignment is Center, expands to Top and Bottom
-	  */
-	def expandingVerticallyAccordingTo(alignment: Alignment) = expandingAccordingTo(Y, alignment)
 }