+++ Reach/src/utopia/reach/component/factory/contextual/ContextualFramedFactory.scala
@@ -4,6 +4,8 @@ import utopia.firmament.context.HasContext
 import utopia.firmament.context.base.BaseContextPropsView
 import utopia.firmament.factory.VariableFramedFactory
 import utopia.firmament.model.enumeration.SizeCategory
+import utopia.paradigm.enumeration.Axis.{X, Y}
+import utopia.paradigm.enumeration.Direction2D.{Down, Up}
 import utopia.paradigm.enumeration.{Axis2D, Direction2D}
 
 /**
@@ -14,6 +16,14 @@ import utopia.paradigm.enumeration.{Axis2D, Direction2D}
   */
 trait ContextualFramedFactory[+Repr] extends VariableFramedFactory[Repr] with HasContext[BaseContextPropsView]
 {
+	def withTop(inset: SizeCategory) = withInset(Up, inset)
+	def withBottom(inset: SizeCategory) = withInset(Down, inset)
+	def withLeft(inset: SizeCategory) = withInset(Direction2D.Left, inset)
+	def withRight(inset: SizeCategory) = withInset(Direction2D.Right, inset)
+	
+	def withHorizontalInsets(inset: SizeCategory) = withInsetsAlong(X, inset)
+	def withVerticalInsets(inset: SizeCategory) = withInsetsAlong(Y, inset)
+	
 	/**
 	  * @param insetSize Size of the insets to apply on all sides
 	  * @return A copy of this factory with the specified symmetric insets